% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_API_results.R
\name{check_API_results}
\alias{check_API_results}
\title{Checks that the result is a 200 status code.}
\usage{
check_API_results(requete)
}
\arguments{
\item{requete}{requete sur l'API d'interet}
}
\value{
un booleen (status = 200 : TRUE / FALSE)
}
\description{
Checks that the result is a 200 status code.
}
\examples{
requete <- httr::GET("http://www.google.com")
check_API_results(requete = requete)
}
