% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_internet.R
\name{check_internet}
\alias{check_internet}
\title{Check if the user has an Internet connexion}
\usage{
check_internet()
}
\value{
a boolean
}
\description{
Check if the user has an Internet connexion
}
\examples{
check_internet()
}
