% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_API_error_code.R
\name{get_API_error_code}
\alias{get_API_error_code}
\title{Permet de retrouver le code erreur d'une requete qui n'a pas fonctionne}
\usage{
get_API_error_code(request)
}
\arguments{
\item{request}{une requete issue d'un appel à un WS xtradata, incorporee
dans un try()}
}
\value{
une liste avec le code d'erreur et le message
}
\description{
Permet de retrouver le code erreur d'une requete qui n'a pas fonctionne
}
\examples{
get_API_error_code(try(xtradata_requete_features(key = "KEY_INCORRECTE", typename = "ST_PARK_P")))
get_API_error_code(try(check_API_results(httr::GET("www.googeul.fr"))))
}
