% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latitude_longitude.R
\name{get_latitude_longitude}
\alias{get_latitude_longitude}
\title{Recupere les données de latitude et longitude à partir d'une requete issue
d'un webservice xtradata}
\usage{
get_latitude_longitude(resultat_requete)
}
\arguments{
\item{resultat_requete}{le data.frame issu d'un appel au WS}
}
\value{
un data frame à 2 colonnes (latitude et longitude)
}
\description{
Recupere les données de latitude et longitude à partir d'une requete issue
d'un webservice xtradata
}
\examples{
\dontrun{
velos <- xtradata_requete_aggregate(key = "Macle",
                                        typename = "PC_CAPTV_P",
                                        rangeStart = "2021-01-01",
                                        rangeEnd = "2021-02-01",
                                        rangeStep = "hour"
                                        )
localisation_velos <- get_latitude_longitude(velos)
}
}
