% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtradata_requete_aggregate.R
\name{xtradata_requete_aggregate}
\alias{xtradata_requete_aggregate}
\title{Renvoie une aggregation des objets historisés sur une période et un pas de temps donné
(service aggregate Xtradata)}
\usage{
xtradata_requete_aggregate(
  key = NULL,
  typename = NULL,
  rangeStart = NULL,
  rangeEnd = NULL,
  rangeStep = NULL,
  rangeFilter = list(hours = 0:23, days = 1:7, publicHolidays = FALSE),
  attributes = NULL,
  filter = NULL
)
}
\arguments{
\item{key}{Cle de connexion (string) (requis)}

\item{typename}{Le nom de la couche (string) (requis)}

\item{rangeStart}{Date / heure de début de la fenêtre de temps.
La valeur sera arrondie à l'heure / au jour ou au mois entamé (selon la valeur du paramètre rangeStep).
Le paramètre accepte une date ou une datetime telle que définie dans la RFC 3339, section 5.6
ou tout format accepté par l'objet JavaScript Date. (date ou datetime) (requis)}

\item{rangeEnd}{Date / heure de fin de la fenêtre de temps.
La valeur sera arrondie à l'heure / au jour ou au mois supérieur à celui entamé (selon la valeur du paramètre rangeStep).
Si aucune valeur n'est précisé, la date courante sera utilisée.
Le paramètre accepte une date ou une datetime telle que définie dans la RFC 3339, section 5.6
ou tout format accepté par l'objet JavaScript Date. (date ou datetime)}

\item{rangeStep}{Type de pas à utiliser pour découper les données.
valeurs autorisées : hour (defaut), day, month (string)}

\item{rangeFilter}{Filtres temporels à appliquer à l'intérieur de la fenêtre de temps
\itemize{
\item hours - Tableau des heure(s) de la journée à retourner, de 0 à 23.
\item days - Tableau des numéros de jours de la semaine à retourner, de 1 à 7, avec 1 pour dimanche et 7 pour samedi.
\item publicHolidays - true pour renvoyer uniquement les occurences des jours fériés. False désactive le filtre
Format liste R ou format JSON (string). Voir exemples
}}

\item{attributes}{Liste des noms des attributs de la couche à retourner en résultat.
Afin d'accélérer les traitements, listez uniquement les attributs que vous sont nécessaires.
Si non précisé, tous les attributs seront retournés.
Format vecteur R ou format array (string). Voir exemples}

\item{filter}{Filtres à appliquer sur les données. Format liste R ou format JSON (string). Voir exemples}
}
\value{
un data frame issu de la requête
}
\description{
Renvoie une aggregation des objets historisés sur une période et un pas de temps donné
(service aggregate Xtradata)
}
\examples{
\dontrun{
# appel sur la couche PC_CAPTE_P
filter <-
list("type" = "BOUCLE",
 "mdate" = list(
      '$gt' = "2020-01-01T08:00:00"
 )
)


filterJSON <-
'{

"type": "BOUCLE",
"mdate": {
  "$gt": "2020-01-01T08:00:00"
 }
}
'

attributes <- c("cdate", "mdate")
attributesArray <- '["cdate", "mdate"]'

# 2 façons d'utiliser le paramètre filter
xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
filter = filter)
xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
filter = filterJSON)

# 2 façons d'utiliser le paramètre attributes
xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
filter = filter, attributes = attributes)
xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
filter = filter, attributes = attributesArray)

# limitation de la requete au 10 premiers resultats
xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
maxfeatures = 10)

}

}
\references{
http://data.bordeaux-metropole.fr/geojson/help/

https://data.bordeaux-metropole.fr/dicopub/#/dico
}
