% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtradata_requete_features.R
\name{xtradata_requete_features}
\alias{xtradata_requete_features}
\title{Renvoie les objets d'une couche (service features Xtradata)}
\usage{
xtradata_requete_features(
  key = NULL,
  typename = NULL,
  crs = "epsg:4326",
  filter = NULL,
  attributes = NULL,
  maxfeatures = NULL,
  showURL = FALSE
)
}
\arguments{
\item{key}{Cle de connexion (string)}

\item{typename}{Le nom de la couche (string)}

\item{crs}{Système de coordonnees de sortie des geometries renvoyees par le serveur GeoJSON
valeurs autorisees : epsg:4326, epsg:3945, epsg:2154, epsg:3857}

\item{filter}{Filtres à appliquer sur les donnees. Format liste R ou format JSON (string). Voir exemples}

\item{attributes}{Liste des noms des attributs de la couche à retourner en resultat.
Afin d'accelerer les traitements, listez uniquement les attributs que vous sont necessaires.
Si non precise, tous les attributs seront retournes.
Format vecteur R ou format array (string). Voir exemples}

\item{maxfeatures}{Nombre maximum d'enregistrement à retourner.
Si 0 ou non precise, tous les enregistrements de la couche seront retournes}

\item{showURL}{afficher l'url interrogee (boolean)}
}
\value{
un data frame issu de la requête
}
\description{
Renvoie les objets d'une couche (service features Xtradata)
}
\references{
http://data.bordeaux-metropole.fr/geojson/help/

https://data.bordeaux-metropole.fr/dicopub/#/dico
}
