% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtradata_requete_features.R
\name{xtradata_requete_features}
\alias{xtradata_requete_features}
\title{Renvoie les objets d'une couche (service features Xtradata)}
\usage{
xtradata_requete_features(
  key = NULL,
  typename = NULL,
  crs = "epsg:4326",
  filter = NULL,
  attributes = NULL,
  maxfeatures = NULL
)
}
\arguments{
\item{key}{Cle de connexion (string)}

\item{typename}{Le nom de la couche (string)}

\item{crs}{Système de coordonnées de sortie des géométries renvoyées par le serveur GeoJSON
valeurs autorisées : epsg:4326, epsg:3945, epsg:2154, epsg:3857}

\item{filter}{Filtres à appliquer sur les données. Format liste R ou format JSON (string). Voir exemples}

\item{attributes}{Liste des noms des attributs de la couche à retourner en résultat.
Afin d'accélérer les traitements, listez uniquement les attributs que vous sont nécessaires.
Si non précisé, tous les attributs seront retournés.
Format vecteur R ou format array (string). Voir exemples}

\item{maxfeatures}{Nombre maximum d'enregistrement à retourner.
Si 0 ou non précisé, tous les enregistrements de la couche seront retournés}
}
\value{
un data frame issu de la requête
}
\description{
Renvoie les objets d'une couche (service features Xtradata)
}
\examples{
\dontrun{
# appel sur la couche PC_CAPTE_P
filter <-
list("type" = "BOUCLE",
 "mdate" = list(
      '$gt' = "2020-01-01T08:00:00"
 )
)
attributes <- c("cdate", "mdate")
filterJSON <-
'{
"type": "BOUCLE",
"mdate": {
  "$gt": "2020-01-01T08:00:00"
 }
}
'
attributesArray <- '["cdate", "mdate"]'
xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
filter = filter)
xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
filter = filterJSON)
xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
filter = filter, attributes = attributes)
xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
filter = filter, attributes = attributesArray)
xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
filter = filter, attributes = attributes, maxfeatures = 10)
xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
filter = filter, attributes = attributesArray, maxfeatures = 10)

}
}
