% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtradata_requete_features.R
\name{xtradata_requete_features}
\alias{xtradata_requete_features}
\title{Renvoie les objets d'une couche (service features Xtradata)}
\usage{
xtradata_requete_features(
  key = NULL,
  typename = NULL,
  crs = "epsg:4326",
  filter = NULL,
  attributes = NULL,
  maxfeatures = NULL
)
}
\arguments{
\item{key}{Cle de connexion (string)}

\item{typename}{Le nom de la couche (string)}

\item{crs}{Système de coordonnees de sortie des geometries renvoyees par le serveur GeoJSON
valeurs autorisees : epsg:4326, epsg:3945, epsg:2154, epsg:3857}

\item{filter}{Filtres à appliquer sur les donnees. Format liste R ou format JSON (string). Voir exemples}

\item{attributes}{Liste des noms des attributs de la couche à retourner en resultat.
Afin d'accelerer les traitements, listez uniquement les attributs que vous sont necessaires.
Si non precise, tous les attributs seront retournes.
Format vecteur R ou format array (string). Voir exemples}

\item{maxfeatures}{Nombre maximum d'enregistrement à retourner.
Si 0 ou non precise, tous les enregistrements de la couche seront retournes}
}
\value{
un data frame issu de la requête
}
\description{
Renvoie les objets d'une couche (service features Xtradata)
}
\examples{
\dontrun{
# appel sur la couche PC_CAPTE_P
filter <- list("type" = "BOUCLE",
               "mdate" = list(
              '$gt' = "2020-01-01T08:00:00"
               )
             )
filterJSON <-
 '{

"type": "BOUCLE",
"mdate": {
  "$gt": "2020-01-01T08:00:00"
 }
}
'
attributes <- list("cdate", "mdate")
attributesArray <- '["cdate", "mdate"]'
# 2 façons d'utiliser le paramètre filter
res1 <- xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
                                 filter = filter)
res2 <- xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
                                 filter = filterJSON)
all.equal(res1, res2)
# 2 façons d'utiliser le paramètre attributes
res3 <- xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
                                 filter = filter, attributes = attributes)
res4 <- xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
       filter = filter, attributes = attributesArray)
all.equal(res3, res4)
# limitation de la requete au 10 premiers resultats
res5 <- xtradata_requete_features(typename  = "PC_CAPTE_P", key = MaCle,
                                 maxfeatures = 10)
nrow(res5)

}

}
\references{
http://data.bordeaux-metropole.fr/geojson/help/

https://data.bordeaux-metropole.fr/dicopub/#/dico
}
